Updated the Android Demo App
