+++ app/src/main/AndroidManifest.xml
@@ -24,6 +24,11 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".CheckoutActivity"
+            android:exported="true"
+            android:theme="@style/Theme.HyperswitchDemoApp">
+        </activity>
     </application>
 
 </manifest>
