+++ app/src/main/res/layout/main_activity.xml
@@ -13,59 +13,15 @@
         android:layout_gravity="center"
         android:orientation="vertical">
 
+
         <Button
-            android:id="@+id/reloadButton"
+            android:id="@+id/checkoutButton"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_margin="20dp"
             android:layout_weight="1"
             android:background="@drawable/rounded_button"
-            android:text="Reload Client Secret"
-            android:textAllCaps="false"
-            android:textColor="#FFFFFF"
-            android:textSize="20sp" />
-
-        <Button
-            android:id="@+id/launchButton"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="20dp"
-            android:layout_weight="1"
-            android:background="@drawable/rounded_button"
-            android:text="Launch Payment Sheet"
-            android:textAllCaps="false"
-            android:textColor="#FFFFFF"
-            android:textSize="20sp" />
-
-        <Button
-            android:id="@+id/launchWebButton"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="20dp"
-            android:layout_weight="1"
-            android:background="@drawable/rounded_button"
-            android:text="Launch Payment Sheet Lite"
-            android:textAllCaps="false"
-            android:textColor="#FFFFFF"
-            android:textSize="20sp" />
-
-        <TextView
-            android:id="@+id/resultText"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="20dp"
-            android:layout_weight="1"
-            android:gravity="center"
-            android:textSize="20sp" />
-
-        <Button
-            android:id="@+id/confirmButton"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="20dp"
-            android:layout_weight="1"
-            android:background="@drawable/rounded_button"
-            android:text="Confirm with Last Used PM"
+            android:text="Checkout"
             android:textAllCaps="false"
             android:textColor="#FFFFFF"
             android:textSize="20sp" />
